+++ Routing/Loader/ConventionalLoader/CollectionGeneratorInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Knp\RadBundle\Routing\Loader\ConventionalLoader;
+
+interface CollectionGeneratorInterface
+{
+    function generateRoute($name, $mapping);
+
+    function generate($name, $mapping);
+}
